# docs/swagger.yaml
basePath: /
definitions:
...
  ark_internal_http_dto.DriverResponse:
    properties:
      driver_id:
        type: string
      license_number:
        type: string
      onboarded_at:
        type: string
      rating:
        type: number
      status:
        type: string
    type: object
...
  ark_internal_httpx.ErrorBody:
    properties:
      error:
        type: string
    type: object
...
  internal_modules_driver.createReq:
    properties:
      license_number:
        type: string
    required:
    - license_number
    type: object
  internal_modules_driver.updateStatusReq:
    properties:
      status:
        enum:
        - available
        - on_trip
        - offline
        type: string
    required:
    - status
    type: object
  internal_modules_driver.updateStatusResponse:
    properties:
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: ZooArk Team
  description: 'ZooArk 叫車平台後端 API。除了 POST /api/users、GET /health 和 GET /readyz 之外，所有
    endpoint 都需要 Firebase Auth ID Token（`Authorization: Bearer <token>`）。金額欄位以 TWD
    分（1/100 元）為單位。'
  title: ZooArk API
  version: 0.1.0
paths:
...
  /api/driver/create:
    post:
      consumes:
      - application/json
      description: Creates the driver profile for the authenticated user (driver_id
        always comes from the Firebase token, never from the body). Starts the driver
        in status "available".
      parameters:
      - description: License info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_driver.createReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ark_internal_http_dto.DriverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Create driver profile
      tags:
      - Driver
  /api/driver/status:
    patch:
      consumes:
      - application/json
      description: Changes the authenticated driver's status to one of available,
        on_trip, offline. Returns the new status echoed back.
      parameters:
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_driver.updateStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_driver.updateStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ark_internal_httpx.ErrorBody'
      security:
      - FirebaseAuth: []
      summary: Update driver status
      tags:
      - Driver
...
schemes:
- http
- https
securityDefinitions:
  FirebaseAuth:
    description: Firebase ID Token, formatted as `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"